Create secrets_dummy.yaml
I added this dummy version of my secrets in order to show how I organize and to show how secrets look for those who may be new to HA or to secrets on HA in general.
This clarifies a lot of the config, which can look confusing without an idea how the secrets.yaml looks.
This file WILL NOT be updated often. If something appears in the config and not in this dummy file, it is because I will not update it often, if ever.